+++ src/main/scala/utils/FixedPointConversion.scala
@@ -29,17 +29,17 @@ object FixedPointConversion {
     BigDecimal.valueOf(scaledToFixed).toBigInt
   }
 
-  def convertFloatMatrixToFixedMatrix(mf: Array[Array[Float]], fixedPointFractionBits: Int, width: Int, signed: Boolean): Array[Array[Int]] = {
-    val m: Array[Array[Int]] = Array.fill(mf.length, mf(0).length)(0)
+  def convertFloatMatrixToFixedMatrix(mf: Array[Array[Float]], fixedPointFractionBits: Int, width: Int, signed: Boolean): Array[Array[BigInt]] = {
+    val m: Array[Array[BigInt]] = Array.fill(mf.length, mf(0).length)(0)
     for (i <- mf.indices) {
       for (j <- mf(0).indices) {
-        m(i)(j) = floatToFixed(mf(i)(j), fixedPointFractionBits, width, signed).toInt
+        m(i)(j) = floatToFixed(mf(i)(j), fixedPointFractionBits, width, signed)
       }
     }
     m
   }
 
-  def convertFixedMatrixToFloatMatrix(m: Array[Array[Int]], fixedPointFractionBits: Int, width: Int, signed: Boolean): Array[Array[Float]] = {
+  def convertFixedMatrixToFloatMatrix(m: Array[Array[BigInt]], fixedPointFractionBits: Int, width: Int, signed: Boolean): Array[Array[Float]] = {
     val mf: Array[Array[Float]] = Array.fill(m.length, m(0).length)(0)
     for (i <- m.indices) {
       for (j <- m(0).indices) {
